+++ src/IncidentCompass.Worker/Program.cs
@@ -1,6 +1,7 @@
 using IncidentCompass.Worker;
 using IncidentCompass.Application;
 using IncidentCompass.Infrastructure;
+using IncidentCompass.Infrastructure.Intake;
 
 var builder = Host.CreateApplicationBuilder(args);
 builder.Services.AddApplication(builder.Configuration);
@@ -8,4 +9,11 @@ builder.Services.AddInfrastructure(builder.Configuration);
 builder.Services.AddWorker(builder.Configuration);
 
 var host = builder.Build();
+var validationExitCode = await TriageConfigurationValidationCommand.RunIfRequestedAsync(args, host.Services);
+if (validationExitCode.HasValue)
+{
+    Environment.ExitCode = validationExitCode.Value;
+    return;
+}
+
 host.Run();